tests: move batch tests from session.rs to batch.rs module
