+++ scylla/tests/integration/session.rs
@@ -35,7 +35,7 @@ use std::sync::{Arc, Mutex};
 use tokio::net::TcpListener;
 use uuid::Uuid;
 
-use scylla::response::query_result::{QueryResult, QueryRowsResult};
+use scylla::response::query_result::QueryResult;
 
 #[tokio::test]
 async fn test_connection_failure() {
@@ -164,200 +164,6 @@ async fn test_unprepared_statement() {
         watchdog += 1;
     }
     assert_eq!(results_from_manual_paging, results);
-}
-
-#[tokio::test]
-async fn test_counter_batch() {
-    use scylla::value::Counter;
-    use scylla_cql::frame::request::batch::BatchType;
-
-    setup_tracing();
-    let session = Arc::new(create_new_session_builder().build().await.unwrap());
-    let ks = unique_keyspace_name();
-
-    // Need to disable tablets in this test because they don't support counters yet.
-    // (https://github.com/scylladb/scylladb/commit/c70f321c6f581357afdf3fd8b4fe8e5c5bb9736e).
-    let mut create_ks = format!("CREATE KEYSPACE IF NOT EXISTS {} WITH REPLICATION = {{'class' : 'NetworkTopologyStrategy', 'replication_factor' : 1}}", ks);
-    if scylla_supports_tablets(&session).await {
-        create_ks += " AND TABLETS = {'enabled': false}"
-    }
-
-    session.ddl(create_ks).await.unwrap();
-    session
-        .ddl(format!(
-            "CREATE TABLE IF NOT EXISTS {}.t_batch (key int PRIMARY KEY, value counter)",
-            ks
-        ))
-        .await
-        .unwrap();
-
-    let statement_str = format!("UPDATE {}.t_batch SET value = value + ? WHERE key = ?", ks);
-    let query = Query::from(statement_str);
-    let prepared = session.prepare(query.clone()).await.unwrap();
-
-    let mut counter_batch = Batch::new(BatchType::Counter);
-    counter_batch.append_statement(query.clone());
-    counter_batch.append_statement(prepared.clone());
-    counter_batch.append_statement(query.clone());
-    counter_batch.append_statement(prepared.clone());
-    counter_batch.append_statement(query.clone());
-    counter_batch.append_statement(prepared.clone());
-
-    // Check that we do not get a server error - the driver
-    // should send a COUNTER batch instead of a LOGGED (default) one.
-    session
-        .batch(
-            &counter_batch,
-            (
-                (Counter(1), 1),
-                (Counter(2), 2),
-                (Counter(3), 3),
-                (Counter(4), 4),
-                (Counter(5), 5),
-                (Counter(6), 6),
-            ),
-        )
-        .await
-        .unwrap();
-}
-
-#[tokio::test]
-async fn test_batch() {
-    setup_tracing();
-    let session = Arc::new(create_new_session_builder().build().await.unwrap());
-    let ks = unique_keyspace_name();
-
-    session.ddl(format!("CREATE KEYSPACE IF NOT EXISTS {} WITH REPLICATION = {{'class' : 'NetworkTopologyStrategy', 'replication_factor' : 1}}", ks)).await.unwrap();
-    session
-        .ddl(format!(
-            "CREATE TABLE IF NOT EXISTS {}.t_batch (a int, b int, c text, primary key (a, b))",
-            ks
-        ))
-        .await
-        .unwrap();
-
-    let prepared_statement = session
-        .prepare(format!(
-            "INSERT INTO {}.t_batch (a, b, c) VALUES (?, ?, ?)",
-            ks
-        ))
-        .await
-        .unwrap();
-
-    // TODO: Add API that supports binding values to statements in batch creation process,
-    // to avoid problem of statements/values count mismatch
-    use scylla::batch::Batch;
-    let mut batch: Batch = Default::default();
-    batch.append_statement(&format!("INSERT INTO {}.t_batch (a, b, c) VALUES (?, ?, ?)", ks)[..]);
-    batch.append_statement(&format!("INSERT INTO {}.t_batch (a, b, c) VALUES (7, 11, '')", ks)[..]);
-    batch.append_statement(prepared_statement.clone());
-
-    let four_value: i32 = 4;
-    let hello_value: String = String::from("hello");
-    let session_clone = session.clone();
-    // We're spawning to a separate task here to test that it works even in that case, because in some scenarios
-    // (specifically if the `BatchValuesIter` associated type is not dropped before await boundaries)
-    // the implicit auto trait propagation on batch will be such that the returned future is not Send (depending on
-    // some lifetime for some unknown reason), so can't be spawned on tokio.
-    // See https://github.com/scylladb/scylla-rust-driver/issues/599 for more details
-    tokio::spawn(async move {
-        let values = (
-            (1_i32, 2_i32, "abc"),
-            (),
-            (1_i32, &four_value, hello_value.as_str()),
-        );
-        session_clone.batch(&batch, values).await.unwrap();
-    })
-    .await
-    .unwrap();
-
-    let mut results: Vec<(i32, i32, String)> = session
-        .query_unpaged(format!("SELECT a, b, c FROM {}.t_batch", ks), &[])
-        .await
-        .unwrap()
-        .into_rows_result()
-        .unwrap()
-        .rows::<(i32, i32, String)>()
-        .unwrap()
-        .collect::<Result<_, _>>()
-        .unwrap();
-
-    results.sort();
-    assert_eq!(
-        results,
-        vec![
-            (1, 2, String::from("abc")),
-            (1, 4, String::from("hello")),
-            (7, 11, String::from(""))
-        ]
-    );
-
-    // Test repreparing statement inside a batch
-    let mut batch: Batch = Default::default();
-    batch.append_statement(prepared_statement);
-    let values = ((4_i32, 20_i32, "foobar"),);
-
-    // This statement flushes the prepared statement cache
-    session
-        .ddl(format!(
-            "ALTER TABLE {}.t_batch WITH gc_grace_seconds = 42",
-            ks
-        ))
-        .await
-        .unwrap();
-    session.batch(&batch, values).await.unwrap();
-
-    let results: Vec<(i32, i32, String)> = session
-        .query_unpaged(
-            format!("SELECT a, b, c FROM {}.t_batch WHERE a = 4", ks),
-            &[],
-        )
-        .await
-        .unwrap()
-        .into_rows_result()
-        .unwrap()
-        .rows::<(i32, i32, String)>()
-        .unwrap()
-        .collect::<Result<_, _>>()
-        .unwrap();
-
-    assert_eq!(results, vec![(4, 20, String::from("foobar"))]);
-}
-
-// This is a regression test for #1134.
-#[tokio::test]
-async fn test_batch_to_multiple_tables() {
-    setup_tracing();
-    let session = create_new_session_builder().build().await.unwrap();
-    let ks = unique_keyspace_name();
-
-    session.ddl(format!("CREATE KEYSPACE IF NOT EXISTS {} WITH REPLICATION = {{'class' : 'NetworkTopologyStrategy', 'replication_factor' : 1}}", ks)).await.unwrap();
-    session.use_keyspace(&ks, true).await.unwrap();
-    session
-        .ddl("CREATE TABLE IF NOT EXISTS t_batch1 (a int, b int, c text, primary key (a, b))")
-        .await
-        .unwrap();
-    session
-        .ddl("CREATE TABLE IF NOT EXISTS t_batch2 (a int, b int, c text, primary key (a, b))")
-        .await
-        .unwrap();
-
-    let prepared_statement = session
-        .prepare(
-            "
-            BEGIN BATCH
-                INSERT INTO t_batch1 (a, b, c) VALUES (?, ?, ?);
-                INSERT INTO t_batch2 (a, b, c) VALUES (?, ?, ?);
-            APPLY BATCH;
-            ",
-        )
-        .await
-        .unwrap();
-
-    session
-        .execute_unpaged(&prepared_statement, (1, 2, "ala", 4, 5, "ma"))
-        .await
-        .unwrap();
 }
 
 #[tokio::test]
@@ -2356,133 +2162,6 @@ async fn test_rate_limit_exceeded_exception() {
         }
         err => panic!("Unexpected error type received: {:?}", err),
     }
-}
-
-// Batches containing LWT queries (IF col = som) return rows with information whether the queries were applied.
-#[tokio::test]
-async fn test_batch_lwts() {
-    let session = create_new_session_builder().build().await.unwrap();
-
-    let ks = unique_keyspace_name();
-    let mut create_ks = format!("CREATE KEYSPACE {} WITH REPLICATION = {{'class': 'NetworkTopologyStrategy', 'replication_factor': 1}}", ks);
-    if scylla_supports_tablets(&session).await {
-        create_ks += " and TABLETS = { 'enabled': false}";
-    }
-    session.ddl(create_ks).await.unwrap();
-    session.use_keyspace(ks.clone(), false).await.unwrap();
-
-    session
-        .ddl("CREATE TABLE tab (p1 int, c1 int, r1 int, r2 int, primary key (p1, c1))")
-        .await
-        .unwrap();
-
-    session
-        .query_unpaged("INSERT INTO tab (p1, c1, r1, r2) VALUES (0, 0, 0, 0)", ())
-        .await
-        .unwrap();
-
-    let mut batch: Batch = Batch::default();
-    batch.append_statement("UPDATE tab SET r2 = 1 WHERE p1 = 0 AND c1 = 0 IF r1 = 0");
-    batch.append_statement("INSERT INTO tab (p1, c1, r1, r2) VALUES (0, 123, 321, 312)");
-    batch.append_statement("UPDATE tab SET r1 = 1 WHERE p1 = 0 AND c1 = 0 IF r2 = 0");
-
-    let batch_res: QueryResult = session.batch(&batch, ((), (), ())).await.unwrap();
-    let batch_deserializer = batch_res.into_rows_result().unwrap();
-
-    // Scylla returns 5 columns, but Cassandra returns only 1
-    let is_scylla: bool = batch_deserializer.column_specs().len() == 5;
-
-    if is_scylla {
-        test_batch_lwts_for_scylla(&session, &batch, &batch_deserializer).await;
-    } else {
-        test_batch_lwts_for_cassandra(&session, &batch, &batch_deserializer).await;
-    }
-}
-
-async fn test_batch_lwts_for_scylla(
-    session: &Session,
-    batch: &Batch,
-    query_rows_result: &QueryRowsResult,
-) {
-    // Alias required by clippy
-    type IntOrNull = Option<i32>;
-
-    // Returned columns are:
-    // [applied], p1, c1, r1, r2
-    let batch_res_rows: Vec<(bool, IntOrNull, IntOrNull, IntOrNull, IntOrNull)> = query_rows_result
-        .rows()
-        .unwrap()
-        .collect::<Result<_, _>>()
-        .unwrap();
-
-    let expected_batch_res_rows = vec![
-        (true, Some(0), Some(0), Some(0), Some(0)),
-        (true, None, None, None, None),
-        (true, Some(0), Some(0), Some(0), Some(0)),
-    ];
-
-    assert_eq!(batch_res_rows, expected_batch_res_rows);
-
-    let prepared_batch: Batch = session.prepare_batch(batch).await.unwrap();
-    let prepared_batch_res: QueryResult =
-        session.batch(&prepared_batch, ((), (), ())).await.unwrap();
-
-    let prepared_batch_res_rows: Vec<(bool, IntOrNull, IntOrNull, IntOrNull, IntOrNull)> =
-        prepared_batch_res
-            .into_rows_result()
-            .unwrap()
-            .rows()
-            .unwrap()
-            .map(|r| r.unwrap())
-            .collect();
-
-    let expected_prepared_batch_res_rows = vec![
-        (false, Some(0), Some(0), Some(1), Some(1)),
-        (false, None, None, None, None),
-        (false, Some(0), Some(0), Some(1), Some(1)),
-    ];
-
-    assert_eq!(prepared_batch_res_rows, expected_prepared_batch_res_rows);
-}
-
-async fn test_batch_lwts_for_cassandra(
-    session: &Session,
-    batch: &Batch,
-    query_rows_result: &QueryRowsResult,
-) {
-    // Alias required by clippy
-    type IntOrNull = Option<i32>;
-
-    // Returned columns are:
-    // [applied]
-    let batch_res_rows: Vec<(bool,)> = query_rows_result
-        .rows()
-        .unwrap()
-        .map(|r| r.unwrap())
-        .collect();
-
-    let expected_batch_res_rows = vec![(true,)];
-
-    assert_eq!(batch_res_rows, expected_batch_res_rows);
-
-    let prepared_batch: Batch = session.prepare_batch(batch).await.unwrap();
-    let prepared_batch_res: QueryResult =
-        session.batch(&prepared_batch, ((), (), ())).await.unwrap();
-
-    // Returned columns are:
-    // [applied], p1, c1, r1, r2
-    let prepared_batch_res_rows: Vec<(bool, IntOrNull, IntOrNull, IntOrNull, IntOrNull)> =
-        prepared_batch_res
-            .into_rows_result()
-            .unwrap()
-            .rows()
-            .unwrap()
-            .map(|r| r.unwrap())
-            .collect();
-
-    let expected_prepared_batch_res_rows = vec![(false, Some(0), Some(0), Some(1), Some(1))];
-
-    assert_eq!(prepared_batch_res_rows, expected_prepared_batch_res_rows);
 }
 
 #[tokio::test]
